pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amm-challenge"
version = "0.1.0"
description = "AMM fee strategy design challenge"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24.0",
    "py-solc-x>=2.0.0",
    "pyrevm>=0.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.scripts]
amm-match = "amm_competition.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["amm_competition*"]
